[StdLib] Refactor ObjC code in Set validation tests
Move tests that need the Objective-C runtime into their own file and remove the
XFAIL on Linux. This partially resolves SR-216.
